requirements.txt
pytest
asyncio
pyrit
allure-pytest